feat: add ticketing system (#11)
* feat: add ticketing system

Added a ticketing system to allow members to create a ticket with staff
These can be used to discuss whatever between staff members

When a ticket is closed, a transcript is generated and sent to a transcript channel

* fix(ci build): haha oopsie

I forgot to include these in the initial commit to this branch

* fix: oopsie doopsie lmao

Fixed port number being incorrect. i was initially going to run this on port 80, but then changed it 3000 and forgot to set the server.listen address

-- additional note
In the attachment saving handler inside of ticketCloseModalHandler, the attachment urls hostname is localhost. This will be changed manually at build time since i didnt want to put the bots IP in the github repo, or make another env variable for it