feat(moderationActions): Now show moderator who banned/kicked a user
We did this to make it a little bit easier to try and determine what moderator did one of these moderationActions
The old behavior: If a member was kicked or banned using the bot, it would say the bot banned them.
New behavior: If a member is banned or kicked using the bot, store it and search for it inside of the leave event. If somethings found, use that. Else default to auditlog or "Unkown User"